+++ .tool-versions
@@ -1,1 +1,1 @@
-rust 1.85.0
+rust 1.79.0
