+++ src/firecracker/Cargo.toml
@@ -35,7 +35,10 @@ vmm-sys-util = { version = "0.12.1", features = ["with-serde"] }
 [dev-dependencies]
 cargo_toml = "0.20.5"
 libc = "0.2.161"
-regex = { version = "1.11.1", default-features = false, features = ["std", "unicode-perl"] }
+regex = { version = "1.11.1", default-features = false, features = [
+  "std",
+  "unicode-perl",
+] }
 
 # Dev-Dependencies for uffd examples
 serde = { version = "1.0.214", features = ["derive"] }
@@ -48,7 +51,12 @@ serde = { version = "1.0.214" }
 serde_json = "1.0.132"
 
 [features]
-tracing = ["log-instrument", "seccompiler/tracing", "utils/tracing", "vmm/tracing"]
+tracing = [
+  "log-instrument",
+  "seccompiler/tracing",
+  "utils/tracing",
+  "vmm/tracing",
+]
 gdb = ["vmm/gdb"]
 
 [lints]
